Bfs, cycle

+++ test/bag_test.c
@@ -14,7 +14,7 @@ void tearDown() {}
 // Helper functions
 void fill_ints() {
   b = bag_init(sizeof(int));
-  for (int i = 0; i < NELEMS(values); i++) {
+  for (size_t i = 0; i < NELEMS(values); i++) {
     bag_add(b, &values[i]);
   }
 }
@@ -22,7 +22,7 @@ void fill_ints() {
 void fill_strs() {
   b = bag_init(sizeof(char *));
   char *str;
-  for (int i = 0; i < NELEMS(strs); i++) {
+  for (size_t i = 0; i < NELEMS(strs); i++) {
     str = strdup(strs[i]);
     bag_add(b, &str);
   }
